Coin pools: BigVector order storage, encoded u128 order ids, kill switch + zero-quote fix

+++ packages/triex/tests/pool_tests.move
@@ -9,7 +9,7 @@
 #[test_only]
 module triex::pool_tests {
     use sui::{object::ID, test_scenario::Scenario};
-    use triex::{order_info::OrderInfo, pool_test_utils};
+    use triex::{coin_order_info::OrderInfo, pool_test_utils};
 
     /// The rates a coin pool resolves at the entry rung, and the cancel-retention
     /// share, re-exported so the integration suites can build expectations
@@ -163,7 +163,7 @@ module triex::pool_tests {
         sender: address,
         pool_id: ID,
         trading_account_id: ID,
-        order_id: u64,
+        order_id: u128,
         test: &mut Scenario,
     ) {
         pool_test_utils::cancel_order<BaseAsset, QuoteAsset>(
@@ -185,7 +185,7 @@ module triex::pool_tests {
         sender: address,
         pool_id: ID,
         trading_account_id: ID,
-        order_id: u64,
+        order_id: u128,
         new_quantity: u64,
         test: &mut Scenario,
     ) {
